validation

+++ controllers/movies.js
@@ -1,6 +1,7 @@
 import Movie from '../models/movie.js';
 import BadRequestError from '../errors/badRequestErr.js';
 import NotFoundError from '../errors/notFoundErr.js';
+import ForbiddenError from '../errors/forbiddenError.js';
 
 export const getMovie = (req, res, next) => {
   Movie.find({})
@@ -10,8 +11,8 @@ export const getMovie = (req, res, next) => {
 
 export const createMovie = (req, res, next) => {
   const movieData = req.body;
-  //const owner = req.user.id;
-  Movie.create({ ...movieData })
+  const owner = req.user.id;
+  Movie.create({ ...movieData, owner })
     .then((movie) => res.send(movie))
     .catch((error) => {
       if (error.name === 'ValidationError') {
@@ -21,13 +22,15 @@ export const createMovie = (req, res, next) => {
     });
 };
 
-// подумать не добавить ли строчку что нельзя удалить фильм чужого пользователя,
-// тогда нужно добавить poplate owner  при создании
 export const deleteMovie = (req, res, next) => {
+  const userId = req.user._id;
   Movie.findById(req.params.movieId)
     .then((movie) => {
       if (!movie) {
         throw new NotFoundError('Фильм с указанным id не найден)');
+      }
+      if (movie.owner.toString() !== userId) {
+        throw new ForbiddenError('Невозможно удалить фильм другого пользователя');
       }
       return Movie.findByIdAndDelete(req.params.movieId);
     })
